+++ src/components/common/index.js
@@ -1,0 +1,4 @@
+//buttons
+//inputs
+//labels
+//cards
