+++ src/components/ui/index.js
@@ -1,0 +1,2 @@
+//Navbar
+//Footer
