+++ src/router/PrivateRoute.jsx
@@ -1,0 +1,3 @@
+export const PrivateRoute = () => {
+  return <div>PrivateRoute</div>;
+};
