+++ src/router/PublicRoute.jsx
@@ -1,0 +1,3 @@
+export const PublicRoute = () => {
+  return <div>PublicRoute</div>;
+};
